pyproject.toml
[tool.poetry]
name = "slimebones.auto"
version = "0.1.0"
description = ""
authors = ["ryzhovalex <thed4rkof@gmail.com>"]
readme = "README.md"
packages = [
    {include = "auto"}
]

[tool.poetry.dependencies]
python = "^3.11"
sbpykit = "^0.4.12"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.8"
coverage = "^7.3.3"
pytest = "^7.4.3"
pytest-asyncio = "^0.23.2"

[tool.poetry.scripts]
auto = "auto.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
